% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_NA.R
\name{check_NA}
\alias{check_NA}
\title{The test of NAs in the data frame.}
\usage{
check_NA(dat)
}
\arguments{
\item{dat}{input dataset. Currently only supports \code{data.frame}. While, \code{data.table} is not supported.}
}
\description{
Sample data to test the package and its functions, is also provided.
This test will check every column in a data.frame for 'NAs'.
In either case (NA exists or not), a message will be printed.
}
