% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_NA.R
\name{find_NA}
\alias{find_NA}
\title{Function to locate instances of NAs (row and column) in sample df
Just for fun ;)}
\usage{
find_NA(dat)
}
\arguments{
\item{dat}{input dataset. Currently only supports \code{data.frame}. While, \code{data.table} is not supported.}
}
\description{
Sample data to test the package and its functions, is also provided
}
