% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_NA.R
\name{fix_NA}
\alias{fix_NA}
\title{Function to replace NAs with blanks
Sample data to test the package and its functions, is also provided
MKE added this comment to check GIT}
\usage{
fix_NA(dat)
}
\arguments{
\item{dat}{input dataset. Currently only supports \code{data.frame}. While, \code{data.table} is not supported.}
}
\description{
Function to replace NAs with blanks
Sample data to test the package and its functions, is also provided
MKE added this comment to check GIT
}
